refactor: reorganize FormRequests into feature-based folders
Restructure FormRequests for better organization:

New structure:
- Bookmarks/ (BookmarkRequest + Create/Delete)
- Likes/ (LikeRequest + Create/Delete)
- Resources/ (Show/Store/Update)
- TechnicalTests/ (Index/Store)
- Users/ (UpdateUserRole)

Changes:
- Updated namespaces to match folder structure
- Fixed authorize() to only check authentication
- Created specific Requests for CRUD operations
- Removed github_id validation (use auth user)
- Commented OldRole dependencies
