+++ app/Http/Requests/Bookmarks/BookmarkRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Bookmarks;
 
 use Illuminate\Foundation\Http\FormRequest;
 use App\Rules\GithubIdRule;
