+++ app/Http/Requests/Likes/CreateLikeRequest.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Http\Requests\Likes;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class CreateLikeRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return $this->user() !== null;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'resource_id' => [
+                'required',
+                'integer',
+                'exists:resources,id'
+            ],
+        ];
+    }
+
+    public function messages(): array
+    {
+        return [
+            'resource_id.required' => 'El ID del recurso es obligatorio.',
+            'resource_id.exists' => 'El recurso no existe.',
+        ];
+    }
+}
