+++ app/Http/Requests/Likes/LikeRequest.php
@@ -2,7 +2,7 @@
 
 declare (strict_types= 1);
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Likes;
 
 use Illuminate\Foundation\Http\FormRequest;
 use App\Rules\GithubIdRule;
