+++ app/Http/Requests/Resources/StoreResourceRequest.php
@@ -2,7 +2,7 @@
 
 declare (strict_types= 1);
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Resources;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
